+++ index.html
@@ -28,5 +28,5 @@
       <button id="start">Start</button>
     </div>
   </body>
-  <script src="scripts/gameboard.js"></script>
+  <script src="scripts/game.js" type="module"></script>
 </html>
