--- scripts/main.js
@@ -1,16 +1,0 @@
-const Game = (player, gameBoard) => {
-
-  const swicthPlayer = () => {
-    
-  }
-
-  const checkMove = () => {
-
-  }
-
-  const checkwinner = () => {
-
-  }
-
-  return { swicthPlayer, checkMove, checkwinner };
-}
